# openapi/api.yaml
openapi: 3.0.3
info:
  title: Edge Cache Lab API
  description: Demo e-commerce API demonstrating CDN and Varnish caching behavior
  version: 0.1.0
  contact:
    name: Edge Cache Lab
servers:
  - url: http://localhost:6081
    description: Local development server
paths:
...
  /product/{id}:
    get:
      operationId: getProduct
      summary: Get product details
      tags:
        - public
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Product ID
      responses:
        '200':
          description: Product details
          headers:
            X-Request-Id:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
            Surrogate-Control:
              schema:
                type: string
            ETag:
              schema:
                type: string
            X-Cache:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/product/{id}:
    post:
      operationId: updateProduct
      summary: Update product (triggers cache purge)
      tags:
        - admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Product ID
        - name: X-Purge-Token
          in: header
          required: true
          schema:
            type: string
          description: Token required to authorize cache purge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Product updated successfully; purge initiated
          headers:
            X-Request-Id:
              schema:
                type: string
            X-Purge-Tags:
              schema:
                type: string
              description: Cache tags purged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          description: Unauthorized; invalid or missing purge token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    Product:
      type: object
      required:
        - id
        - name
        - slug
        - price
        - inStock
      properties:
        id:
          type: string
          example: "prod-001"
        name:
          type: string
          example: "Wireless Headphones"
        slug:
          type: string
          example: "wireless-headphones"
        description:
          type: string
          example: "High-quality wireless headphones with noise cancellation"
        price:
          type: number
          format: float
          example: 199.99
        currency:
          type: string
          example: "USD"
        inStock:
          type: boolean
          example: true
        categoryId:
          type: string
          example: "cat-001"
        imageUrl:
          type: string
          example: "https://example.com/images/prod-001.jpg"
        tags:
          type: array
          items:
            type: string
          example: ["electronics", "audio", "wireless"]
...
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        inStock:
          type: boolean
...
    Error:
      type: object
      required:
        - error
        - message
        - requestId
      properties:
        error:
          type: string
          example: "not_found"
        message:
          type: string
          example: "Product not found"
        requestId:
          type: string
          example: "req-abc123"
        timestamp:
          type: string
          format: date-time
          example: "2026-02-13T17:23:00Z"